extract util function toUntyped in interop module
