+++ src/model.js
@@ -11,7 +11,3 @@ export const FormArray = NgFormArray;
 export const FormControl = NgFormControl;
 export const FormGroup = NgFormGroup;
 export const FormBuilder =  NgFormBuilder;
-
-export function toUntyped(control) {
-  return control;
-}
