+++ src/public_api.ts
@@ -1,2 +1,3 @@
 export * from './model';
 export * from './module';
+export * from './interop';
