+++ test/util.ts
@@ -1,6 +1,7 @@
-import {FormControl, FormGroup, toUntyped} from '../src/model';
+import {FormControl,} from '../src/model';
 import {Foo} from './interfaces';
 import {FormControl as AngularFormControl} from '@angular/forms';
+import {toUntyped} from '../src/interop';
 
 const formGroup = new FormControl<Foo>();
